requirements.txt
# This file defines the image's dependencies
Flask
pillow
keras
numpy
pandas